requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.10
python-dotenv==0.21.1
gunicorn==21.2.0
